[Docs] Add docs to deploy project on Astro Cloud
Add documentation for deploying the project in the /dev directory on Astro Cloud. I created a Makefile target called "deploy" for this purpose.

For this exercise, I set up an Astro deployment on AWS Cloud and a RayCluster on GKE.

To deploy on Astro Cloud, run the command make deploy.

closes: https://github.com/astronomer/astro-provider-ray/issues/88