Fix and document local development setup
Previously, we were only able to run the example DAG , with a few tweaks. None of the remaining example DAGs would run in the local Kind cluster when using the local Astro CLI setup.

There were multiple reasons for this, including network limitations, default Kind limitations and lack of configuration on the Docker and Kind side.

This PR shares the pain points and the fixes, allowing anyone running the Ray Provider exmaple DAGs in MacOS (M1) to be able to run the example DAGs locally and successfully spin up Ray Clusters from Airflow.