Fix: convert headers from connection from string to dictionary (#124)
The headers are retrieved from the connection as a string. But they are expected to be a dictionary. Getting the error:

```
ERROR - Failed during execution with error: Failed to create Ray JobSubmissionClient: headers must be a dict, got <class 'str'>
```

The fix is to use json.loads to convert the header to dict before assigning them to the headers attribute.

Fixes: #119